+++ opengl/oglplus/include/oglplus/prog_var/wrapper.hpp
@@ -101,7 +101,7 @@ public:
 	{
 		return ProgVar(
 			ProgramName(this->_program),
-			this->_location+offset
+			this->_location+static_cast<GLuint>(offset)
 		);
 	}
